+++ src/formats/formats.ts
@@ -1,6 +1,4 @@
-/* eslint-disable no-control-regex */
-import { isAsciiIdn, isUri, isIdnEmail, isIri, isIriReference, isIdn } from "@hyperjump/json-schema-formats";
-import validUrl from "valid-url";
+import { isWebUri } from "valid-url";
 import { getTypeOf } from "../utils/getTypeOf";
 import { JsonSchemaValidatorParams, ValidationReturnType } from "../Keyword";
 import settings from "../settings";
@@ -13,14 +11,8 @@ const matchDate = /^(\d\d\d\d)-(\d\d)-(\d\d)$/;
 const matchTime =
     /^(?<time>(?:([0-1]\d|2[0-3]):[0-5]\d:(?<second>[0-5]\d|60)))(?:\.\d+)?(?<offset>(?:z|[+-]([0-1]\d|2[0-3])(?::?[0-5]\d)?))$/i;
 const DAYS = [0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
 const isValidJsonPointer = /^(?:\/(?:[^~/]|~0|~1)*)*$/;
 const isValidRelativeJsonPointer = /^(?:0|[1-9][0-9]*)(?:#|(?:\/(?:[^~/]|~0|~1)*)*)$/;
-const isValidURIRef =
-    /^(?:[a-z][a-z0-9+\-.]*:)?(?:\/?\/(?:(?:[a-z0-9\-._~!$&'()*+,;=:]|%[0-9a-f]{2})*@)?(?:\[(?:(?:(?:(?:[0-9a-f]{1,4}:){6}|::(?:[0-9a-f]{1,4}:){5}|(?:[0-9a-f]{1,4})?::(?:[0-9a-f]{1,4}:){4}|(?:(?:[0-9a-f]{1,4}:){0,1}[0-9a-f]{1,4})?::(?:[0-9a-f]{1,4}:){3}|(?:(?:[0-9a-f]{1,4}:){0,2}[0-9a-f]{1,4})?::(?:[0-9a-f]{1,4}:){2}|(?:(?:[0-9a-f]{1,4}:){0,3}[0-9a-f]{1,4})?::[0-9a-f]{1,4}:|(?:(?:[0-9a-f]{1,4}:){0,4}[0-9a-f]{1,4})?::)(?:[0-9a-f]{1,4}:[0-9a-f]{1,4}|(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?))|(?:(?:[0-9a-f]{1,4}:){0,5}[0-9a-f]{1,4})?::[0-9a-f]{1,4}|(?:(?:[0-9a-f]{1,4}:){0,6}[0-9a-f]{1,4})?::)|[Vv][0-9a-f]+\.[a-z0-9\-._~!$&'()*+,;=:]+)\]|(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?)|(?:[a-z0-9\-._~!$&'"()*+,;=]|%[0-9a-f]{2})*)(?::\d*)?(?:\/(?:[a-z0-9\-._~!$&'"()*+,;=:@]|%[0-9a-f]{2})*)*|\/(?:(?:[a-z0-9\-._~!$&'"()*+,;=:@]|%[0-9a-f]{2})+(?:\/(?:[a-z0-9\-._~!$&'"()*+,;=:@]|%[0-9a-f]{2})*)*)?|(?:[a-z0-9\-._~!$&'"()*+,;=:@]|%[0-9a-f]{2})+(?:\/(?:[a-z0-9\-._~!$&'"()*+,;=:@]|%[0-9a-f]{2})*)*)?(?:\?(?:[a-z0-9\-._~!$&'"()*+,;=:@/?]|%[0-9a-f]{2})*)?(?:#(?:[a-z0-9\-._~!$&'"()*+,;=:@/?]|%[0-9a-f]{2})*)?$/i;
-// uri-template: https://tools.ietf.org/html/rfc6570
-const isValidURITemplate =
-    /^(?:(?:[^\x00-\x20"'<>%\\^`{|}]|%[0-9a-f]{2})|\{[+#./;?&=,!@|]?(?:[a-z0-9_]|%[0-9a-f]{2})+(?::[1-9][0-9]{0,3}|\*)?(?:,(?:[a-z0-9_]|%[0-9a-f]{2})+(?::[1-9][0-9]{0,3}|\*)?)*\})*$/i;
 const isValidDurationString = /^P(?!$)(\d+Y)?(\d+M)?(\d+W)?(\d+D)?(T(?=\d)(\d+H)?(\d+M)?(\d+S)?)?$/;
 
 // Default Json-Schema formats: date-time, email, hostname, ipv4, ipv6, uri, uriref
@@ -80,6 +72,7 @@ export const formats: Record<string, (options: JsonSchemaValidatorParams) => Val
             });
         }
     },
+
     email: ({ node, pointer, data }) => {
         const { schema } = node;
         if (typeof data !== "string" || data === "") {
@@ -125,78 +118,6 @@ export const formats: Record<string, (options: JsonSchemaValidatorParams) => Val
             return node.createError("format-email-error", { value: data, pointer, schema });
         }
         return undefined;
-    },
-
-    hostname: ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || isAsciiIdn(data)) {
-            return undefined;
-        }
-        return node.createError("format-hostname-error", { value: data, pointer, schema });
-    },
-
-    "idn-hostname": ({ node, pointer, data }) => {
-        if (typeof data !== "string" || isIdn(data)) {
-            return undefined;
-        }
-        return node.createError("format-idn-hostname-error", { value: data, pointer, schema: node.schema });
-    },
-
-    /**
-     * @draft 7
-     * [RFC6531] https://json-schema.org/draft-07/json-schema-validation.html#RFC6531
-     */
-    "idn-email": ({ node, pointer, data }) => {
-        if (typeof data !== "string" || data === "" || isIdnEmail(data)) {
-            return undefined;
-        }
-        return node.createError("format-email-error", { value: data, pointer, schema: node.schema });
-    },
-
-    ipv4: ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "") {
-            return undefined;
-        }
-        if (data && data[0] === "0") {
-            // leading zeroes should be rejected, as they are treated as octals
-            return node.createError("format-ipv4-leading-zero-error", { value: data, pointer, schema });
-        }
-        if (data.length <= 15 && isValidIPV4.test(data)) {
-            return undefined;
-        }
-        return node.createError("format-ipv4-error", { value: data, pointer, schema });
-    },
-
-    ipv6: ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "") {
-            return undefined;
-        }
-        if (data && data[0] === "0") {
-            // leading zeroes should be rejected, as they are treated as octals
-            return node.createError("format-ipv6-leading-zero-error", { value: data, pointer, schema });
-        }
-        if (data.length <= 45 && isValidIPV6.test(data)) {
-            return undefined;
-        }
-        return node.createError("format-ipv6-error", { value: data, pointer, schema });
-    },
-
-    iri: ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "" || isIri(data)) {
-            return undefined;
-        }
-        return node.createError("format-iri-error", { value: data, pointer, schema });
-    },
-
-    "iri-reference": ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "" || isIriReference(data)) {
-            return undefined;
-        }
-        return node.createError("format-iri-reference-error", { value: data, pointer, schema });
     },
 
     "json-pointer": ({ node, pointer, data }) => {
@@ -291,38 +212,9 @@ export const formats: Record<string, (options: JsonSchemaValidatorParams) => Val
         return undefined;
     },
 
-    uri: ({ node, pointer, data }) => {
-        if (typeof data !== "string" || data === "" || isUri(data)) {
-            return undefined;
-        }
-        return node.createError("format-uri-error", { value: data, pointer, schema: node.schema });
-    },
-
-    "uri-reference": ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "") {
-            return undefined;
-        }
-        if (isValidURIRef.test(data)) {
-            return undefined;
-        }
-        return node.createError("format-uri-reference-error", { value: data, pointer, schema });
-    },
-
-    "uri-template": ({ node, pointer, data }) => {
-        const { schema } = node;
-        if (typeof data !== "string" || data === "") {
-            return undefined;
-        }
-        if (isValidURITemplate.test(data)) {
-            return undefined;
-        }
-        return node.createError("format-uri-template-error", { value: data, pointer, schema });
-    },
-
     url: ({ node, data, pointer }) => {
         const { schema } = node;
-        if (data === "" || validUrl.isWebUri(`${data}`)) {
+        if (data === "" || isWebUri(`${data}`)) {
             return undefined;
         }
         return node.createError("format-url-error", { value: data, pointer, schema });